修改类型

+++ include/time/cooling.h
@@ -7,32 +7,31 @@
 class CCooling {
 private:
 protected:
-    uint64_t m_uBeginTick;    /*冷却开始时间*/
-    uint64_t m_uEndTick;      /*冷却结束时间*/
-    uint8_t  m_state;          //状态
+    uint64_t m_uBeginTick; /*冷却开始时间*/
+    uint64_t m_uEndTick; /*冷却结束时间*/
+    uint8_t m_state; //状态
 public:
     CCooling();
 
-    virtual~CCooling() { }
+    virtual ~CCooling() { }
 
 public:
     void clearCool();
 
     /*--->[ 获得冷却时间 ]*/
-    uint32_t getCoolTick();
+    uint64_t getCoolTick();
 
-    uint32_t getPassTick();
+    uint64_t getPassTick();
 
     /*--->[ 获得总冷却时间 ]*/
-    uint32_t getTotalTick();
+    uint64_t getTotalTick();
 
     /*--->[ 开始冷却 ]*/
-    bool beginCooling(uint32_t uTick);
+    bool beginCooling(uint64_t uTick);
 
     bool isTimeOut();
 
     void setState(uint8_t state);
 
     uint8_t getState();
-
 };
